T1: extract FlightPicker into pure view + thin authed container (paid Duffel order stays container-only)
